ISPN-2192 Allow using Distributed Execution against any cache (not just clustered caches)
